Add some colouring for the RISC OS types.
Types were only coloured by the name of the files, from the extension.
As RISC OS files don't have extensions, this left a lot of white.
So here we have a patch that forces the colours to change according
to the RISC OS types and some of the paths used in Pyromaniac.